+++ src/containers/scenes/mobile/nrc-landing-scene-mobile/nrc-landing-scene-mobile-constants.js
